+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## 1.1.0 - Unreleased
 
+### Added
+
+* Support for the 'Basic weather maps' API
+* Missing 'ground level pressure' and 'sea level pressure' measurements to CurrentWeather
+
 ### Changed
 
 * Instead of wrapping API results and exceptions in an ApiResponse, return or throw them directly.